Add description for 'break' key
The `E0 46` scan code is used for the 'break' key. It is usually used
in combination with 'ctrl' to interrupt a console application.

Note that the 'pause' key, which 'break' usually shares a physical key
with on traditional 107-key keyboards, uses a different scan code. Those
keyboards would report report the `E0 46` scancode only when that
physical key is pressed together with 'ctrl' or 'shift'.