Added the remapping I use for Surface keyboards
Load this SKL file in SharpKeys v3.7 or higher to flip the functionality of some of the function keys.  For me, I leave Fn-Lock on so the F1-F7 keys act like function keys while Home, End, PgUp, and PgDn act as if Fn is off.

The key list is:
F9 to Home
F10 to End
F11 to PgUp
F12 to PgDn
Home to F9
End to F10
PgUp to F11
PgDn to F12
Right Alt to Left Windows
Caps Lock to disabled

Remember to save the new key mappings to the Registry and reboot for the changes take effect!